Refactor: 상품 페이지 코드 모듈화 #118
