+++ src/constants/index.ts
@@ -1,31 +1,29 @@
-const _id = localStorage.getItem('_id');
-
 export const DASHBOARD_MENU = {
   buyer: [
     {
       id: 1,
       title: '구매자 대시보드',
-      url: `/user/${_id}`,
+      url: '/user',
     },
     {
       id: 2,
       title: '내 정보 수정',
-      url: `/user/${_id}/buyer-info`,
+      url: 'info',
     },
     {
       id: 3,
       title: '내 주문 내역',
-      url: `/user/${_id}/buyer-orderlist`,
+      url: 'order-list',
     },
     {
       id: 4,
       title: '찜한 상품',
-      url: `/user/${_id}/buyer-favorite`,
+      url: 'favorites',
     },
     {
       id: 5,
       title: '최근 본 상품',
-      url: `/user/${_id}/recently-viewed-items`,
+      url: 'recently-viewed',
     },
   ],
 
@@ -33,17 +31,17 @@ export const DASHBOARD_MENU = {
     {
       id: 1,
       title: '판매 상품 관리',
-      url: `/user/${_id}/product-manager`,
+      url: 'seller/products',
     },
     {
       id: 2,
       title: '판매자 정보',
-      url: `/user/${_id}/seller-info`,
+      url: 'seller/info',
     },
     {
       id: 3,
       title: '주문 관리',
-      url: `/user/${_id}/seller-order-manager`,
+      url: 'seller/orders',
     },
   ],
 };
