+++ src/pages/product/SearchPage.tsx
@@ -10,7 +10,7 @@ import RefreshIcon from '@mui/icons-material/Refresh';
 
 import { SearchSection } from '../../components/search/SearchSection';
 import { useSearchStore, useRecentViewProductStore } from '../../lib/store';
-import StickyNavbar from '../../components/NavigationBar';
+import StickyNavbar from '../../components/navbar/NavigationBar';
 import { useSort } from '../../hooks/useSort';
 import { useEffect, useState } from 'react';
 import { IProduct } from '../../type';
@@ -22,7 +22,7 @@ import {
 } from '../../constants';
 import { useFetchProducts } from '../../hooks/useFetchProducts';
 import { ProductGrid } from '../../components/search/ProductGrid';
-import { useScrollToTop } from '../../hooks/useScrollToTop';
+import MobileNavBar from '../../components/navbar/MobileNavBar';
 
 export function SearchPage() {
   const { searchResult, setSearchResult } = useSearchStore();
@@ -36,7 +36,6 @@ export function SearchPage() {
   const [selectedPrice, setSelectedPrice] = useState('전체');
   const [selectedShippingFee, setSelectedShippingFee] = useState('전체');
   const [isDataFetched, setIsDataFetched] = useState(false);
-  useScrollToTop();
 
   const { addRecentViewProduct } = useRecentViewProductStore() as {
     addRecentViewProduct: Function;
@@ -276,6 +275,7 @@ export function SearchPage() {
           </Grid>
         </Grid>
       </Box>
+      <MobileNavBar />
     </Box>
   );
 }
